requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
APScheduler==3.6.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.8
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-jsonfield-backport==1.0.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.4
frozenlist==1.3.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
psycopg==3.1.2
psycopg-binary==3.1.2
psycopg2==2.9.3
pycodestyle==2.9.1
PyJWT==2.5.0
pyparsing==3.0.9
python-environ==0.4.54
python-monkey-business==1.0.0
python-telegram-bot==13.14
pytils==0.4.1
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
typing_extensions==4.3.0
tzdata==2022.5
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
yarl==1.8.1
